requirements.txt
google-search-results
dotenv
pillow
wikipedia
